requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
Flask==2.1.2
flask-accept==0.0.6
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.0
